fully funtioning website

+++ fetch-utils.js
@@ -20,10 +20,9 @@ export async function createTodo(todo) {
 
 export async function deleteAllTodos() {
     // delete all todos for this user in supabase
-    const response = await client
-        .from('todos')
-        .delete()
-        .match({ user_id: client.auth.user().id, id: id });
+    const response = await client.from('todos').delete().match({ user_id: client.auth.user().id });
+    // eslint-disable-next-line no-console
+    console.log(response);
     // once you have a response from supabase, comment this back in:
     return checkError(response);
 }
